requirements.txt
imapclient
python-dotenv
claude-agent-sdk
